Remove flow dependent logic outside of the design-carousel component
