+++ client/landing/stepper/declarative-flow/internals/steps-repository/design-carousel/index.tsx
@@ -1,7 +1,10 @@
 import { Design } from '@automattic/design-picker';
-import { StepContainer } from '@automattic/onboarding';
+import { useLocale } from '@automattic/i18n-utils';
+import { ECOMMERCE_FLOW, StepContainer, isLinkInBioFlow } from '@automattic/onboarding';
+import { useMediaQuery } from '@wordpress/compose';
 import { useDispatch } from '@wordpress/data';
 import { useI18n } from '@wordpress/react-i18n';
+import { StarterDesigns, useStarterDesignsQuery } from 'calypso/../packages/data-stores/src';
 import AsyncLoad from 'calypso/components/async-load';
 import FormattedHeader from 'calypso/components/formatted-header';
 import { ONBOARD_STORE } from 'calypso/landing/stepper/stores';
@@ -9,10 +12,89 @@ import { recordTracksEvent } from 'calypso/lib/analytics/tracks';
 import type { Step } from '../../types';
 import './style.scss';
 
+const shouldOnlyDisplayMobileCarousel = ( flow: string | null | undefined ) => {
+	switch ( true ) {
+		case isLinkInBioFlow( flow ):
+			return true;
+		default:
+			return false;
+	}
+};
+
+const getEcommerceDesigns = ( allDesigns: StarterDesigns ) => {
+	let selectedDesigns = allDesigns?.designs;
+	const selectedDesignSlugs = [ 'tsubaki', 'amulet', 'tazza', 'zaino', 'thriving-artist' ];
+
+	// If we have a restricted set of designs, filter out all unwanted designs
+	const filteredDesigns = selectedDesigns.filter( ( design ) =>
+		selectedDesignSlugs.includes( design.slug )
+	);
+
+	// Now order the filtered set based on the supplied slugs.
+	selectedDesigns = selectedDesignSlugs
+		.map( ( selectedDesignSlug ) =>
+			filteredDesigns.find( ( design ) => design.slug === selectedDesignSlug )
+		)
+		.filter( ( selectedDesign ) => !! selectedDesign ) as Design[];
+
+	return selectedDesigns;
+};
+
+const getLinkInBioDesigns = ( allDesigns: StarterDesigns ) => {
+	const designs =
+		allDesigns?.designs.filter(
+			( design ) =>
+				design.is_virtual &&
+				design.categories.some( ( category ) => category.slug === 'link-in-bio' )
+		) ?? [];
+
+	return designs;
+};
+
+const getCarouselDesktopOptions = (
+	flow: string | null | undefined,
+	isLargerThan1440px: boolean
+) => {
+	switch ( true ) {
+		case flow === ECOMMERCE_FLOW:
+			return {
+				w: isLargerThan1440px ? 1920 : 1280,
+				vpw: 1920,
+				vph: 1280,
+				format: 'png',
+			};
+		default:
+			return;
+	}
+};
+
+const getFlowDesigns = (
+	allDesigns: StarterDesigns | undefined,
+	flow: string | null | undefined
+) => {
+	if ( ! allDesigns ) {
+		return null;
+	}
+
+	switch ( true ) {
+		case isLinkInBioFlow( flow ):
+			return getLinkInBioDesigns( allDesigns );
+		case flow === ECOMMERCE_FLOW:
+			return getEcommerceDesigns( allDesigns );
+		default:
+			return allDesigns?.designs;
+	}
+};
+
 const DesignCarousel: Step = function DesignCarousel( { navigation, flow } ) {
 	const { goNext, goBack, submit } = navigation;
 	const { __ } = useI18n();
+	const locale = useLocale();
 	const { setSelectedDesign } = useDispatch( ONBOARD_STORE );
+	const isLargerThan1440px = useMediaQuery( '(min-width: 1440px)' );
+	const { data: allDesigns } = useStarterDesignsQuery( {
+		_locale: locale,
+	} );
 
 	function pickDesign( _selectedDesign: Design ) {
 		setSelectedDesign( _selectedDesign );
@@ -31,8 +113,9 @@ const DesignCarousel: Step = function DesignCarousel( { navigation, flow } ) {
 					require="@automattic/design-carousel"
 					placeholder={ null }
 					onPick={ pickDesign }
-					flow={ flow }
-					selectedDesignSlugs={ [ 'tsubaki', 'amulet', 'tazza', 'zaino', 'thriving-artist' ] }
+					selectedDesigns={ getFlowDesigns( allDesigns, flow ) }
+					onlyDisplayMobileCarousel={ shouldOnlyDisplayMobileCarousel( flow ) }
+					carouselDesktopOptions={ getCarouselDesktopOptions( flow, isLargerThan1440px ) }
 				/>
 			}
 			recordTracksEvent={ recordTracksEvent }
